requirements.txt
discord.py==2.3.2
discord-racket==0.0.14
